Refactored to use src instead of stash
