+++ src/settings.py
@@ -1,0 +1,1 @@
+path = ".src"
