--- stash/settings.py
@@ -1,1 +1,0 @@
-path = ".stash"
